requirements.txt
alabaster==0.7.12
alembic==1.4.2
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.3
argh==0.26.2
asgiref==3.2.5
asn1crypto==1.3.0
astroid==2.1.0
astropy==4.0
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
autopep8==1.5
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.3
bokeh==2.0.0
bootstrap-admin==0.4.4
Bootstrap-Flask==1.2.0
boto==2.49.0
Bottleneck==1.3.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
conda==4.8.3
conda-build==3.19.1
conda-package-handling==1.6.0
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.12.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.12.0
Django==3.0.3
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-extensions==3.0.8
django-haystack==3.0b2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-mptt==0.11.0
django-redis==4.12.1
django-werkzeug-debugger-runserver==0.3.1
django-widget-tweaks==1.4.8
docutils==0.16
dominate==2.5.1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.10
flake8==3.7.9
Flask==1.1.1
Flask-CKEditor==0.4.3
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
flask-mongoengine==0.9.5
Flask-Share==0.1.1
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
fsspec==0.6.3
future==0.18.2
geohash2==1.1
gevent==1.4.0
glob2==0.7
greenlet==0.4.15
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==17.3.1
hypothesis==5.7.0
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jieba==0.42.1
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
libarchive-c==2.9
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
mongoengine==0.19.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
mysqlclient==1.3.14
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.4
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parsel==1.5.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
Protego==0.1.16
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.5
PyDispatcher==2.0.5
pydocstyle==5.0.2
pyecharts==1.7.1
pyflakes==2.1.1
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.2.2
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.4.1
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.1
pytest-runner==5.2
python-dateutil==2.8.1
python-editor==1.0.4
python-geohash==0.8.5
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python3-openid==3.2.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.1
QtPy==1.9.0
queuelib==1.5.0
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
Rtree==0.9.4
ruamel-yaml==0.15.80
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==1.6.0
seaborn==0.10.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.17.0
singledispatch==3.4.0.3
sip==4.19.20
six==1.12.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.4
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.1.2
spyder==4.1.1
spyder-kernels==1.9.0
SQLAlchemy==1.3.15
sqlparse==0.3.1
statsmodels==0.11.1
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.7
visitor==0.1.3
w3lib==1.20.0
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
Whoosh==2.7.4
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.1
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwings==0.18.0
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.29.0
zict==2.0.0
zipp==3.1.0
zope.interface==5.0.0
